Updated notebook code to do some inference and plotting with trained model.
